refactor(api): decompose MCP tool registration
Move MCP authentication, policy, serialization, and domain registration into
one-way modules with explicit ownership. The server now constructs the
transport, registers resources, and delegates tool registration.

Update tests to patch canonical owners and guard the public tool inventory
without compatibility facades or module mutation.
