+++ apps/api/src/sibyl/mcp_tools/__init__.py
@@ -1,0 +1,1 @@
+"""MCP tool domains and their shared support modules."""
