+++ apps/api/src/sibyl/server.py
@@ -1,1906 +1,10 @@
-"""MCP server definition using MCPServer with streamable HTTP transport.
+"""MCP server construction and resource registration."""
 
-Exposes 13 tools and 2 resources:
-- Tools: search, explore, context, expand_neighbors, fetch_slice, remember,
-  reflect, add, manage, logs, synthesis_plan, synthesis_draft, synthesis_verify
-- Resources: sibyl://health, sibyl://stats
-"""
-
-from collections.abc import Awaitable, Callable, Iterable, Mapping
-from copy import deepcopy
-from dataclasses import asdict, dataclass
-from functools import wraps
-from typing import Annotated, Any, Literal, cast
-from uuid import UUID, uuid4
-
-import structlog
 from mcp.server import MCPServer
-from mcp.server.auth.middleware.auth_context import get_access_token
-from pydantic import Field
 
-from sibyl.api.context_audit import log_context_pack_audit, log_reflection_audit
-from sibyl.api.idempotency import (
-    complete_idempotency_record,
-    idempotency_lock,
-    idempotency_record_pending,
-    idempotency_request_hash,
-    reserve_idempotency_record,
-)
-from sibyl.auth.api_key_common import api_key_memory_scope_key
-from sibyl.auth.mcp_auth import (
-    effective_api_key_scopes,
-    insufficient_mcp_scope_message,
-    mcp_scopes_allow,
-)
-from sibyl.auth.memory_targets import RelationshipReaderScope, validate_relationship_targets
+import sibyl.mcp_tools.context as mcp_context
 from sibyl.config import settings
-from sibyl.persistence.auth_runtime import (
-    authenticate_api_key,
-    create_project_record,
-    has_owner_membership,
-    log_memory_audit_event,
-    resolve_accessible_project_graph_ids,
-    resolve_accessible_team_scope_keys,
-    resolve_org_role,
-)
-from sibyl.persistence.content_common import ApiIdempotencyRecord
-from sibyl.services.recall_limits import (
-    RecallConcurrencyLimitExceededError,
-    recall_concurrency_slot,
-)
-from sibyl.services.work_item_workflow import WorkItemAction
-from sibyl_core.auth.context import MemoryPolicyContext
-from sibyl_core.auth.memory_policy import (
-    MemoryPolicyAction,
-    MemoryPolicyDecision,
-    authorize_memory_write,
-    server_provenance_metadata,
-    stamp_memory_scope_metadata,
-)
-from sibyl_core.memory_pipeline.capture import MemoryCaptureRequest, MemoryCaptureService
-from sibyl_core.models.relations import DECLARABLE_PREDICATE_HELP
-from sibyl_core.services.surreal_content import (
-    MemoryScope,
-    get_raw_memory,
-    get_raw_memory_by_source_id,
-)
-from sibyl_core.tools.context import DEFAULT_MARKDOWN_TOKEN_BUDGET
-from sibyl_core.tools.traverse import (
-    DEFAULT_EXPAND_LIMIT,
-    DEFAULT_NEIGHBOR_CONTENT_MAX_CHARS,
-    DEFAULT_SLICE_CONTENT_MAX_CHARS,
-    DEFAULT_SLICE_WINDOW,
-    DEFAULT_TRAVERSAL_DEPTH,
-)
-
-log = structlog.get_logger()
-
-MemoryKind = Literal[
-    "episode",
-    "decision",
-    "plan",
-    "idea",
-    "claim",
-    "artifact",
-    "procedure",
-    "domain",
-    "session",
-    "pattern",
-    "rule",
-]
-SynthesisOutputKind = Literal[
-    "documentation",
-    "report",
-    "briefing",
-    "roadmap",
-    "release_notes",
-    "audit_packet",
-    "handbook",
-    "custom",
-]
-SynthesisDepthKind = Literal["brief", "standard", "deep"]
-# The tool schema, not just the prose description, has to carry the predicate
-# vocabulary: a model picks parameter values from the schema it is handed.
-DeclaredRelatedTo = Annotated[list[str] | None, Field(description=DECLARABLE_PREDICATE_HELP)]
-SynthesisArtifactKind = Literal["markdown", "json"]
-
-MCP_ENTITY_PROJECT_POLICY_ACTIONS = {
-    "add_note",
-    "archive_epic",
-    "archive_task",
-    "block_task",
-    "complete_epic",
-    "complete_task",
-    "estimate",
-    "start_epic",
-    "start_task",
-    "submit_review",
-    "suggest",
-    "unblock_task",
-    "update_epic",
-    "update_task",
-}
-MCP_PROJECT_ID_POLICY_ACTIONS = {"detect_cycles", "prioritize"}
-
-
-@dataclass(frozen=True)
-class McpContext:
-    """Context extracted from MCP authentication token."""
-
-    org_id: str
-    user_id: str | None = None
-    scopes: list[str] | None = None
-    # API key project restrictions (None = all, list = only these)
-    api_key_project_ids: list[str] | None = None
-    api_key_memory_space_ids: list[str] | None = None
-    api_key_memory_scope_keys: list[str] | None = None
-    org_role: str | None = None
-    delegated_authority: str | None = None
-    agent_id: str | None = None
-    # True when the caller authenticated with an API key. Scope enforcement is
-    # an API-key concern on both surfaces: user sessions carry no scope claim,
-    # and REST gates scopes only on its own API-key branch.
-    is_api_key: bool = False
-
-    def to_memory_policy_context(
-        self,
-        *,
-        memory_space: str | None = None,
-        scope_key: str | None = None,
-        project_id: str | None = None,
-        accessible_projects: Iterable[str] | None = None,
-        accessible_teams: Iterable[str] | None = None,
-        accessible_delegations: Iterable[str] | None = None,
-        source_surface: str = "mcp",
-    ) -> MemoryPolicyContext:
-        return MemoryPolicyContext(
-            actor_user_id=self.user_id,
-            organization_id=self.org_id,
-            organization_role=self.org_role,
-            accessible_projects=frozenset(str(value) for value in accessible_projects)
-            if accessible_projects is not None
-            else None,
-            accessible_teams=frozenset(str(value) for value in accessible_teams)
-            if accessible_teams is not None
-            else None,
-            accessible_delegations=frozenset(str(value) for value in accessible_delegations)
-            if accessible_delegations is not None
-            else None,
-            delegated_authority=self.delegated_authority,
-            agent_id=self.agent_id,
-            project_id=project_id,
-            memory_space=memory_space,
-            scope_key=scope_key,
-            source_surface=source_surface,
-        )
-
-
-async def _get_mcp_context() -> McpContext | None:
-    """Extract full context (org_id, user_id, scopes) from MCP token.
-
-    Returns:
-        McpContext if authenticated, None otherwise.
-    """
-    token = get_access_token()
-    if token is None:
-        return None
-
-    raw = token.token
-    if not raw:
-        return None
-
-    # API Key authentication
-    if raw.startswith("sk_"):
-        auth = await authenticate_api_key(raw)
-        if auth:
-            # Convert project UUIDs to graph IDs (strings)
-            project_ids = (
-                [str(pid) for pid in auth.project_ids] if auth.project_ids is not None else None
-            )
-            org_id = str(auth.organization_id)
-            user_id = str(auth.user_id)
-            org_role = await resolve_org_role(org_id=org_id, user_id=user_id)
-            return McpContext(
-                org_id=org_id,
-                user_id=user_id,
-                scopes=auth.scopes,
-                api_key_project_ids=project_ids,
-                api_key_memory_space_ids=[
-                    str(memory_space_id)
-                    for memory_space_id in getattr(auth, "memory_space_ids", None) or []
-                ]
-                if getattr(auth, "memory_space_ids", None) is not None
-                else None,
-                api_key_memory_scope_keys=[
-                    memory_space.policy_key
-                    for memory_space in getattr(auth, "memory_spaces", None) or []
-                ]
-                if getattr(auth, "memory_spaces", None) is not None
-                else None,
-                org_role=org_role,
-                is_api_key=True,
-            )
-        return None
-
-    # JWT authentication
-    from sibyl.auth.jwt import JwtError, verify_access_token
-
-    try:
-        claims = verify_access_token(raw)
-    except JwtError:
-        return None
-
-    org_id = claims.get("org")
-    user_id = claims.get("sub")
-
-    if org_id:
-        log.debug("mcp_context", org_id=org_id, user_id=user_id)
-        org_role = await resolve_org_role(
-            org_id=str(org_id),
-            user_id=str(user_id) if user_id else None,
-        )
-        return McpContext(
-            org_id=str(org_id),
-            user_id=str(user_id) if user_id else None,
-            scopes=claims.get("scopes"),
-            org_role=org_role,
-        )
-    return None
-
-
-async def _optional_mcp_org_id() -> str | None:
-    """Extract the organization ID when the caller presents a usable credential.
-
-    Health reporting works without an org, so an anonymous caller gets None
-    rather than an error. A credential that is presented still has to clear the
-    read gate: failing scopes are refused, not silently downgraded to anonymous.
-
-    Returns:
-        The organization ID string if authenticated and org-scoped, None otherwise.
-    """
-    ctx = await _get_mcp_context()
-    if ctx is None:
-        return None
-    _authorize_mcp_scope(ctx, write=False)
-    return ctx.org_id
-
-
-def _authorize_mcp_scope(ctx: McpContext, *, write: bool) -> None:
-    """Enforce API-key scopes for an MCP tool call.
-
-    Raises:
-        ValueError: If the key lacks the scope this tool requires.
-    """
-    if not ctx.is_api_key:
-        return
-    scopes = effective_api_key_scopes(ctx.scopes)
-    if mcp_scopes_allow(scopes, write=write):
-        return
-    log.warning(
-        "mcp_insufficient_scope",
-        org_id=ctx.org_id,
-        write=write,
-        scopes=sorted(scopes),
-    )
-    raise ValueError(insufficient_mcp_scope_message(scopes, write=write))
-
-
-async def _require_mcp_context(*, write: bool = False) -> McpContext:
-    """Require full MCP context including user_id.
-
-    Args:
-        write: True when the caller is about to mutate state, which demands the
-            `api:write` scope from any key that carries granular REST scopes.
-
-    Raises:
-        ValueError: If no context is available or its scopes are insufficient.
-
-    Returns:
-        McpContext with org_id and user_id.
-    """
-    ctx = await _get_mcp_context()
-    if not ctx:
-        raise ValueError("Organization context required. Authenticate with an org-scoped token.")
-    _authorize_mcp_scope(ctx, write=write)
-    if write and ctx.org_role not in {"owner", "admin", "member"}:
-        raise ValueError("organization_write_forbidden")
-    return ctx
-
-
-def _serialize_mcp_idempotency[**P, R](
-    path: str,
-    *,
-    action_scoped: bool = False,
-) -> Callable[[Callable[P, Awaitable[R]]], Callable[P, Awaitable[R]]]:
-    def decorator(func: Callable[P, Awaitable[R]]) -> Callable[P, Awaitable[R]]:
-        @wraps(func)
-        async def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
-            raw_key: object = kwargs.get("idempotency_key")
-            data = kwargs.get("data")
-            if raw_key is None and isinstance(data, Mapping):
-                raw_key = data.get("idempotency_key")
-            if not isinstance(raw_key, str) or not raw_key.strip():
-                return await func(*args, **kwargs)
-
-            scoped_path = path
-            if action_scoped:
-                action = str(kwargs.get("action", "unknown")).lower().strip()
-                scoped_path = f"{path}/{action}"
-            ctx = await _require_mcp_context(write=True)
-            async with idempotency_lock(
-                organization_id=ctx.org_id,
-                principal_id=ctx.user_id or "unknown",
-                method="MCP",
-                path=scoped_path,
-                key=raw_key.strip(),
-            ):
-                return await func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
-
-
-async def _get_accessible_projects(ctx: McpContext) -> set[str] | None:
-    """Get project IDs the user can access based on their permissions.
-
-    Combines user permissions with API key project restrictions (if any).
-
-    Returns:
-        Set of accessible project graph IDs, or None if no filtering needed (admin).
-    """
-    if not ctx.user_id:
-        # No user context - can't filter by user permissions
-        # But still enforce API key restrictions if present
-        if ctx.api_key_project_ids is not None:
-            return set(ctx.api_key_project_ids)
-        return None
-
-    return await resolve_accessible_project_graph_ids(
-        user_id=ctx.user_id,
-        org_id=ctx.org_id,
-        scopes=ctx.scopes,
-        api_key_project_ids=ctx.api_key_project_ids,
-    )
-
-
-async def _resolve_mcp_project_scope(
-    ctx: McpContext,
-    project: str | None,
-    *,
-    require_project_when_restricted: bool = False,
-) -> set[str] | None:
-    """Resolve accessible project scope for MCP tools."""
-    accessible_projects = await _get_accessible_projects(ctx)
-    if accessible_projects is None:
-        if project:
-            return {project}
-        return None
-    if project:
-        if project not in accessible_projects:
-            raise ValueError(f"Project access denied: {project}")
-        return {project}
-    if require_project_when_restricted:
-        raise ValueError("Project is required when MCP credentials are project-scoped.")
-    return accessible_projects
-
-
-async def _compile_mcp_context_pack(
-    *,
-    goal: str,
-    intent: Literal["build", "plan", "ideate", "research", "debug", "decide", "learn", "general"],
-    layer: Literal["wake", "recall", "deep_search"],
-    domain: str | None,
-    project: str | None,
-    agent_id: str | None,
-    limit: int,
-    include_related: bool,
-    related_limit: int,
-    audit: bool = False,
-    markdown_token_budget: int | None = DEFAULT_MARKDOWN_TOKEN_BUDGET,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import (
-        compile_context as _compile_context,
-        context_pack_to_dict,
-        context_pack_to_markdown,
-    )
-
-    ctx = await _require_mcp_context()
-    accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-    memory_scope = "project" if project else "private"
-    scope_key = project
-    if not _mcp_context_pack_scope_allowed(
-        ctx,
-        project=project,
-        accessible_projects=accessible_projects,
-    ):
-        _deny_mcp_api_key_memory_scope(
-            ctx=ctx,
-            action=MemoryPolicyAction.READ,
-            memory_scope=memory_scope,
-            scope_key=scope_key,
-            surface="mcp_context",
-        )
-    if ctx.user_id is None:
-        raise ValueError("User context required for recall.")
-    try:
-        async with recall_concurrency_slot(
-            organization_id=ctx.org_id,
-            user_id=ctx.user_id,
-            organization_role=ctx.org_role,
-        ):
-            pack = await _compile_context(
-                goal=goal,
-                intent=intent,
-                layer=layer,
-                domain=domain,
-                project=project,
-                accessible_projects=accessible_projects,
-                principal_id=ctx.user_id,
-                agent_id=agent_id,
-                limit=limit,
-                include_related=include_related,
-                related_limit=related_limit,
-                audit=audit,
-                organization_id=ctx.org_id,
-                allowed_memory_scope_keys=set(ctx.api_key_memory_scope_keys)
-                if ctx.api_key_memory_scope_keys is not None
-                else None,
-            )
-    except RecallConcurrencyLimitExceededError as exc:
-        raise ValueError("recall_concurrency_limit_exceeded") from exc
-    payload = context_pack_to_dict(pack)
-    payload["markdown"] = context_pack_to_markdown(pack, token_budget=markdown_token_budget)
-    await log_context_pack_audit(
-        user_id=ctx.user_id,
-        organization_id=ctx.org_id,
-        pack=pack,
-        project=project,
-        accessible_projects=accessible_projects,
-        source_surface="mcp_context",
-        agent_id=agent_id,
-        limit=limit,
-        include_related=include_related,
-        related_limit=related_limit,
-    )
-    return payload
-
-
-def _log_mcp_policy_decision(
-    *,
-    ctx: McpContext,
-    decision: MemoryPolicyDecision,
-    surface: str,
-) -> None:
-    log.info(
-        "mcp_memory_policy_decision",
-        action=decision.action.value,
-        allowed=decision.allowed,
-        memory_scope=decision.memory_scope.value,
-        organization_id=ctx.org_id,
-        policy_reason=decision.reason,
-        principal_id=ctx.user_id,
-        scope_key=decision.scope_key,
-        surface=surface,
-    )
-
-
-def _mcp_memory_scope_allowed(ctx: McpContext, *, memory_scope: str, scope_key: str | None) -> bool:
-    allowed = ctx.api_key_memory_scope_keys
-    if allowed is None:
-        return True
-    effective_scope_key = ctx.user_id if memory_scope == MemoryScope.PRIVATE.value else scope_key
-    return api_key_memory_scope_key(memory_scope, effective_scope_key) in set(allowed)
-
-
-def _mcp_context_pack_scope_allowed(
-    ctx: McpContext,
-    *,
-    project: str | None,
-    accessible_projects: set[str] | None,
-) -> bool:
-    if project:
-        return _mcp_memory_scope_allowed(
-            ctx, memory_scope=MemoryScope.PROJECT.value, scope_key=project
-        )
-    allowed = ctx.api_key_memory_scope_keys
-    if allowed is None:
-        return True
-    allowed_keys = set(allowed)
-    if _mcp_memory_scope_allowed(ctx, memory_scope=MemoryScope.PRIVATE.value, scope_key=None):
-        return True
-    if accessible_projects is None:
-        return False
-    return any(
-        api_key_memory_scope_key(MemoryScope.PROJECT.value, project_id) in allowed_keys
-        for project_id in accessible_projects
-    )
-
-
-def _deny_mcp_api_key_memory_scope(
-    *,
-    ctx: McpContext,
-    action: MemoryPolicyAction,
-    memory_scope: str,
-    scope_key: str | None,
-    surface: str,
-) -> None:
-    try:
-        normalized_scope = MemoryScope(memory_scope)
-    except ValueError:
-        normalized_scope = MemoryScope.PRIVATE
-    decision = MemoryPolicyDecision(
-        action=action,
-        allowed=False,
-        reason="api_key_memory_space_denied",
-        memory_scope=normalized_scope,
-        scope_key=scope_key,
-    )
-    _log_mcp_policy_decision(ctx=ctx, decision=decision, surface=surface)
-    raise ValueError(decision.reason)
-
-
-def _authorize_mcp_memory_write(
-    *,
-    ctx: McpContext,
-    memory_scope: str,
-    scope_key: str | None,
-    accessible_projects: set[str] | None,
-    surface: str,
-    accessible_teams: set[str] | None = None,
-) -> MemoryPolicyDecision:
-    policy_context = ctx.to_memory_policy_context(
-        memory_space=memory_scope,
-        scope_key=scope_key,
-        project_id=scope_key,
-        accessible_projects=accessible_projects,
-        accessible_teams=accessible_teams,
-        source_surface=surface,
-    )
-    decision = authorize_memory_write(
-        policy_context=policy_context,
-    )
-    _log_mcp_policy_decision(ctx=ctx, decision=decision, surface=surface)
-    if not decision.allowed:
-        raise ValueError(decision.reason)
-    if not _mcp_memory_scope_allowed(ctx, memory_scope=memory_scope, scope_key=scope_key):
-        _deny_mcp_api_key_memory_scope(
-            ctx=ctx,
-            action=MemoryPolicyAction.WRITE,
-            memory_scope=memory_scope,
-            scope_key=scope_key,
-            surface=surface,
-        )
-    return decision
-
-
-def _append_unique_ids(existing: list[str] | None, additions: list[str] | None) -> list[str] | None:
-    links = list(existing or [])
-    seen = set(links)
-    for item in additions or []:
-        if item not in seen:
-            links.append(item)
-            seen.add(item)
-    return links or None
-
-
-async def _resolve_mcp_capture_links(
-    *,
-    ctx: McpContext,
-    project: str | None,
-    related_to: list[str] | None,
-    task_ids: list[str] | None,
-    active_task: bool,
-    accessible_projects: set[str] | None,
-) -> list[str] | None:
-    links = _append_unique_ids(related_to, task_ids)
-    if not active_task or not project:
-        return links
-
-    from sibyl_core.tools.core import explore
-
-    try:
-        response = await explore(
-            mode="list",
-            types=["task"],
-            project=project,
-            status="doing",
-            limit=2,
-            organization_id=ctx.org_id,
-            principal_id=ctx.user_id,
-            accessible_projects=accessible_projects,
-            allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-        )
-    except Exception as exc:
-        log.warning("mcp_active_task_lookup_failed", project=project, error=str(exc))
-        return links
-
-    entities = getattr(response, "entities", [])
-    if len(entities) != 1:
-        return links
-
-    task_id = getattr(entities[0], "id", None)
-    if not task_id:
-        return links
-
-    return _append_unique_ids(links, [str(task_id)])
-
-
-async def _validate_mcp_relationship_targets(
-    *,
-    ctx: McpContext,
-    related_to: list[str] | None,
-    accessible_projects: set[str] | None,
-) -> None:
-    if not related_to:
-        return
-    from sibyl_core.services.graph import get_surreal_graph_runtime
-
-    runtime = await get_surreal_graph_runtime(ctx.org_id)
-    await validate_relationship_targets(
-        entity_manager=runtime.entity_manager,
-        related_to=related_to,
-        scope=RelationshipReaderScope.from_values(
-            user_id=ctx.user_id,
-            accessible_projects=accessible_projects,
-            memory_grants=ctx.api_key_memory_scope_keys,
-        ),
-    )
-
-
-@_serialize_mcp_idempotency("mcp/remember")
-async def _remember_mcp_memory(
-    *,
-    title: str,
-    content: str,
-    kind: MemoryKind,
-    domain: str | None,
-    project: str | None,
-    tags: list[str] | None,
-    related_to: list[str] | None,
-    task_ids: list[str] | None = None,
-    active_task: bool = True,
-    metadata: dict[str, Any] | None = None,
-    retrieval_keys: list[str] | None = None,
-    idempotency_key: str | None = None,
-    spans: list[dict[str, Any]] | None = None,
-    atomic: bool = False,
-    probes: list[str] | None = None,
-) -> dict[str, Any]:
-    from sibyl_core.memory_pipeline.structure import build_memory_structure
-    from sibyl_core.services.surreal_content import remember_raw_memory
-    from sibyl_core.tools.core import add
-
-    ctx = await _require_mcp_context(write=True)
-    accessible_projects = await _resolve_mcp_project_scope(
-        ctx,
-        project,
-        require_project_when_restricted=True,
-    )
-    if not ctx.user_id:
-        raise ValueError("User context required to remember raw source material.")
-    if idempotency_key is not None:
-        idempotency_key = idempotency_key.strip()
-        if not idempotency_key or len(idempotency_key) > 255:
-            raise ValueError("idempotency_key must be a non-empty string up to 255 characters")
-
-    # Ahead of the idempotency claim: a refused plan must not consume the key,
-    # or the caller's corrected retry would replay the rejection.
-    build_memory_structure(content.strip(), spans=spans, atomic=atomic, probes=probes)
-
-    memory_scope = "project" if project else "private"
-    write_decision = _authorize_mcp_memory_write(
-        ctx=ctx,
-        memory_scope=memory_scope,
-        scope_key=project,
-        accessible_projects=accessible_projects,
-        surface="mcp_remember",
-    )
-    resolved_links = await _resolve_mcp_capture_links(
-        ctx=ctx,
-        project=project,
-        related_to=related_to,
-        task_ids=task_ids,
-        active_task=active_task,
-        accessible_projects=accessible_projects,
-    )
-    await _validate_mcp_relationship_targets(
-        ctx=ctx,
-        related_to=resolved_links,
-        accessible_projects=accessible_projects,
-    )
-
-    idempotency_payload = {
-        "title": title,
-        "content": content,
-        "kind": kind,
-        "domain": domain,
-        "project": project,
-        "tags": tags,
-        "related_to": related_to,
-        "task_ids": task_ids,
-        "active_task": active_task,
-        "metadata": metadata,
-        "retrieval_keys": retrieval_keys,
-        "spans": spans,
-        "atomic": atomic,
-        "probes": probes,
-    }
-    idempotency_path = "mcp/remember"
-    idempotency_claim: ApiIdempotencyRecord | None = None
-    if idempotency_key is not None:
-        record, claimed = await reserve_idempotency_record(
-            organization_id=UUID(ctx.org_id),
-            principal_id=ctx.user_id,
-            idempotency_key=idempotency_key,
-            method="MCP",
-            path=idempotency_path,
-            payload=idempotency_payload,
-            content_session=None,
-        )
-        if not claimed:
-            if record.request_hash != idempotency_request_hash(idempotency_payload):
-                raise ValueError("idempotency_key was already used for a different request")
-            if idempotency_record_pending(record):
-                # This branch runs under _serialize_mcp_idempotency's lock, so
-                # no executor is live on this key: the reservation was
-                # interrupted before its receipt completed. Adopt it and
-                # re-execute; completion overwrites the same record id.
-                log.warning(
-                    "mcp_idempotency_interrupted_takeover",
-                    path=idempotency_path,
-                    organization_id=ctx.org_id,
-                )
-                idempotency_claim = record
-            else:
-                replayed = cast("dict[str, Any]", deepcopy(record.response_body))
-                receipt = replayed.get("mutation_receipt")
-                if isinstance(receipt, dict):
-                    replayed["mutation_receipt"] = {**receipt, "replayed": True}
-                return replayed
-        else:
-            idempotency_claim = record
-
-    full_metadata = dict(metadata or {})
-    full_metadata["capture_kind"] = kind
-    full_metadata["organization_id"] = ctx.org_id
-    if domain:
-        full_metadata["domain"] = domain
-    if project:
-        full_metadata["project_id"] = project
-    if ctx.user_id:
-        full_metadata["created_by"] = ctx.user_id
-    capture_request = MemoryCaptureRequest(
-        title=title,
-        content=content,
-        entity_type=kind,
-        domain=domain,
-        tags=tags,
-        related_to=resolved_links,
-        metadata=full_metadata,
-        retrieval_keys=retrieval_keys,
-        provenance={"remember_kind": kind, "related_to": resolved_links or []},
-        source_id=f"mcp:remember:{kind}",
-        memory_scope=memory_scope,
-        scope_key=project,
-        principal_id=ctx.user_id,
-        capture_surface="mcp",
-        spans=spans,
-        atomic=atomic,
-        probes=probes,
-    )
-    raw_revision = 1
-
-    async def remember_raw(
-        request: MemoryCaptureRequest,
-    ) -> Mapping[str, Any]:
-        nonlocal raw_revision
-        raw_memory = await remember_raw_memory(
-            organization_id=ctx.org_id,
-            principal_id=ctx.user_id,
-            source_id=request.source_id,
-            raw_content=request.content,
-            title=request.title,
-            memory_scope=request.memory_scope,
-            scope_key=request.scope_key,
-            tags=list(request.tags) if request.tags is not None else None,
-            metadata=dict(request.metadata),
-            provenance=dict(request.provenance),
-            capture_surface=request.capture_surface,
-        )
-        raw_revision = getattr(raw_memory, "revision", 1)
-        return {"id": raw_memory.id, "source_id": raw_memory.source_id}
-
-    async def create_graph_entity(
-        request: MemoryCaptureRequest,
-        graph_metadata: Mapping[str, Any],
-    ) -> Mapping[str, Any]:
-        result = await add(
-            title=request.title,
-            content=request.content,
-            entity_type=request.entity_type,
-            category=request.domain,
-            tags=list(request.tags) if request.tags is not None else None,
-            related_to=list(request.related_to) if request.related_to is not None else None,
-            metadata=dict(graph_metadata),
-            # The graph writer strips server-owned keys from `metadata`,
-            # because that bag is caller input everywhere else it is used. The
-            # capture pipeline stamped these two from the raw write it just
-            # completed, so they travel in the argument that survives the
-            # strip; without it the row reaches the graph naming no capture,
-            # and both the correction write-through and the projection
-            # boundary lose the only link back.
-            capture_provenance=server_provenance_metadata(graph_metadata),
-            project=project,
-            memory_scope=request.memory_scope,
-            scope_key=request.scope_key,
-            principal_id=request.principal_id,
-            accessible_projects=accessible_projects,
-            allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-            retrieval_keys=list(request.retrieval_keys)
-            if request.retrieval_keys is not None
-            else None,
-            spans=list(request.spans) if request.spans is not None else None,
-            atomic=request.atomic,
-            probes=list(request.probes) if request.probes is not None else None,
-        )
-        return _to_dict(result)
-
-    capture_result = await MemoryCaptureService(
-        remember_raw_memory=remember_raw,
-        create_graph_entity=create_graph_entity,
-    ).capture(capture_request)
-    payload = capture_result.to_payload()
-    if capture_result.raw_policy_reason is None:
-        payload.pop("raw_policy_reason", None)
-    payload["policy_reason"] = write_decision.reason
-    affected_records = (
-        [f"raw_captures:{capture_result.raw_memory_id}"] if capture_result.raw_memory_id else []
-    )
-    payload["mutation_receipt"] = _mcp_mutation_receipt(
-        {"idempotency_key": idempotency_key},
-        applied=True,
-        revision=raw_revision,
-        affected_records=affected_records,
-    )
-    if idempotency_claim is not None:
-        await complete_idempotency_record(
-            idempotency_claim,
-            response_status_code=200,
-            response_body=payload,
-            content_session=None,
-        )
-    return payload
-
-
-async def _reflect_mcp_memory(
-    *,
-    content: str,
-    source_title: str = "Session reflection",
-    intent: Literal[
-        "build", "plan", "ideate", "research", "debug", "decide", "learn", "general"
-    ] = "general",
-    domain: str | None = None,
-    project: str | None = None,
-    related_to: list[str] | None = None,
-    task_ids: list[str] | None = None,
-    active_task: bool = True,
-    persist: bool = False,
-    persist_source: bool = True,
-    persist_review: bool = False,
-    cited_ids: list[str] | str | None = None,
-    limit: int = 12,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import (
-        reflect_memory,
-        reflection_pack_to_dict,
-        reflection_pack_to_markdown,
-    )
-
-    ctx = await _require_mcp_context(write=True)
-    accessible_projects = await _resolve_mcp_project_scope(
-        ctx,
-        project,
-        require_project_when_restricted=persist,
-    )
-    resolved_links = await _resolve_mcp_capture_links(
-        ctx=ctx,
-        project=project,
-        related_to=related_to,
-        task_ids=task_ids,
-        active_task=active_task and persist,
-        accessible_projects=accessible_projects,
-    )
-    memory_scope = "project" if project else "private"
-    scope_key = project
-    if persist:
-        _authorize_mcp_memory_write(
-            ctx=ctx,
-            memory_scope=memory_scope,
-            scope_key=scope_key,
-            accessible_projects=accessible_projects,
-            surface="mcp_reflect",
-        )
-        await _validate_mcp_relationship_targets(
-            ctx=ctx,
-            related_to=resolved_links,
-            accessible_projects=accessible_projects,
-        )
-    pack = await reflect_memory(
-        content=content,
-        source_title=source_title,
-        intent=intent,
-        domain=domain,
-        project=project,
-        related_to=resolved_links,
-        organization_id=ctx.org_id,
-        principal_id=ctx.user_id,
-        accessible_projects=accessible_projects,
-        memory_scope=memory_scope,
-        scope_key=scope_key,
-        persist=persist,
-        persist_source=persist_source,
-        persist_review=persist_review,
-        limit=limit,
-    )
-    payload = reflection_pack_to_dict(pack)
-    payload["markdown"] = reflection_pack_to_markdown(pack)
-    if cited_ids:
-        from sibyl_core.tools.usage_citation import record_cited_item_usages
-
-        try:
-            payload["citation_usage"] = await record_cited_item_usages(
-                cited_ids,
-                organization_id=ctx.org_id,
-                principal_id=ctx.user_id,
-                project_id=project,
-                source_surface="mcp_reflect",
-                request_metadata={
-                    "active_task": active_task,
-                    "intent": intent,
-                    "persist": persist,
-                    "source_title": source_title,
-                },
-            )
-        except Exception as exc:
-            log.warning(
-                "mcp_reflect_citation_usage_failed",
-                error_type=type(exc).__name__,
-                exc_info=True,
-            )
-    await log_reflection_audit(
-        user_id=ctx.user_id,
-        organization_id=ctx.org_id,
-        pack=pack,
-        project=project,
-        accessible_projects=accessible_projects,
-        source_surface="mcp_reflect",
-        persist=persist,
-        persist_source=persist_source,
-        persist_review=persist_review,
-        active_task=active_task,
-        related_to=resolved_links,
-        task_ids=task_ids,
-        limit=limit,
-    )
-    return payload
-
-
-async def _synthesis_mcp_plan(
-    *,
-    goal: str,
-    output_type: SynthesisOutputKind = "documentation",
-    audience: str | None = None,
-    depth: SynthesisDepthKind = "standard",
-    seed_query: str | None = None,
-    project: str | None = None,
-    domain: str | None = None,
-    entity_ids: list[str] | None = None,
-    decision_ids: list[str] | None = None,
-    task_ids: list[str] | None = None,
-    artifact_ids: list[str] | None = None,
-    required_sections: list[dict[str, Any] | str] | None = None,
-    constraints: list[str] | None = None,
-    max_sections: int = 6,
-    include_neighborhoods: bool = True,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import synthesis_plan
-
-    ctx = await _require_mcp_context()
-    accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-    return await synthesis_plan(
-        goal=goal,
-        output_type=output_type,
-        audience=audience,
-        depth=depth,
-        seed_query=seed_query,
-        project=project,
-        domain=domain,
-        entity_ids=entity_ids,
-        decision_ids=decision_ids,
-        task_ids=task_ids,
-        artifact_ids=artifact_ids,
-        required_sections=required_sections,
-        constraints=constraints,
-        max_sections=max_sections,
-        include_neighborhoods=include_neighborhoods,
-        organization_id=ctx.org_id,
-        principal_id=ctx.user_id,
-        accessible_projects=accessible_projects,
-        allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-    )
-
-
-async def _synthesis_mcp_verify(
-    *,
-    goal: str,
-    output_type: SynthesisOutputKind = "documentation",
-    audience: str | None = None,
-    depth: SynthesisDepthKind = "standard",
-    seed_query: str | None = None,
-    project: str | None = None,
-    domain: str | None = None,
-    entity_ids: list[str] | None = None,
-    decision_ids: list[str] | None = None,
-    task_ids: list[str] | None = None,
-    artifact_ids: list[str] | None = None,
-    required_sections: list[dict[str, Any] | str] | None = None,
-    constraints: list[str] | None = None,
-    max_sections: int = 6,
-    include_neighborhoods: bool = True,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import synthesis_verify
-
-    ctx = await _require_mcp_context()
-    accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-    return await synthesis_verify(
-        goal=goal,
-        output_type=output_type,
-        audience=audience,
-        depth=depth,
-        seed_query=seed_query,
-        project=project,
-        domain=domain,
-        entity_ids=entity_ids,
-        decision_ids=decision_ids,
-        task_ids=task_ids,
-        artifact_ids=artifact_ids,
-        required_sections=required_sections,
-        constraints=constraints,
-        max_sections=max_sections,
-        include_neighborhoods=include_neighborhoods,
-        organization_id=ctx.org_id,
-        principal_id=ctx.user_id,
-        accessible_projects=accessible_projects,
-        allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-    )
-
-
-async def _synthesis_mcp_draft(
-    *,
-    goal: str,
-    output_type: SynthesisOutputKind = "documentation",
-    audience: str | None = None,
-    depth: SynthesisDepthKind = "standard",
-    seed_query: str | None = None,
-    project: str | None = None,
-    domain: str | None = None,
-    entity_ids: list[str] | None = None,
-    decision_ids: list[str] | None = None,
-    task_ids: list[str] | None = None,
-    artifact_ids: list[str] | None = None,
-    required_sections: list[dict[str, Any] | str] | None = None,
-    constraints: list[str] | None = None,
-    max_sections: int = 6,
-    include_neighborhoods: bool = True,
-    output_format: SynthesisArtifactKind = "markdown",
-    remember: bool = False,
-    memory_scope: str = "private",
-    scope_key: str | None = None,
-    tags: list[str] | None = None,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import synthesis_draft
-
-    ctx = await _require_mcp_context(write=remember)
-    accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-    resolved_scope_key = scope_key
-    policy_reason: str | None = None
-    if remember:
-        write_accessible_projects = accessible_projects
-        if memory_scope == "project":
-            resolved_scope_key = resolved_scope_key or project
-            write_accessible_projects = await _resolve_mcp_project_scope(
-                ctx,
-                resolved_scope_key,
-                require_project_when_restricted=True,
-            )
-        decision = _authorize_mcp_memory_write(
-            ctx=ctx,
-            memory_scope=memory_scope,
-            scope_key=resolved_scope_key,
-            accessible_projects=write_accessible_projects,
-            surface="mcp_synthesis",
-        )
-        policy_reason = decision.reason
-
-    payload = await synthesis_draft(
-        goal=goal,
-        output_type=output_type,
-        audience=audience,
-        depth=depth,
-        seed_query=seed_query,
-        project=project,
-        domain=domain,
-        entity_ids=entity_ids,
-        decision_ids=decision_ids,
-        task_ids=task_ids,
-        artifact_ids=artifact_ids,
-        required_sections=required_sections,
-        constraints=constraints,
-        max_sections=max_sections,
-        include_neighborhoods=include_neighborhoods,
-        output_format=output_format,
-        remember=remember,
-        memory_scope=memory_scope,
-        scope_key=resolved_scope_key,
-        tags=tags,
-        organization_id=ctx.org_id,
-        principal_id=ctx.user_id,
-        accessible_projects=accessible_projects,
-        allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-    )
-    if policy_reason:
-        payload["policy_reason"] = policy_reason
-    return payload
-
-
-# Work items are addressed by project membership through project_id, which
-# every read path already enforces. Giving them a memory scope as well would
-# add a second audience channel that only MCP-created rows carry, splitting
-# the population in two and hiding them from lookups that thread project
-# access but no reader identity.
-_UNSCOPED_ENTITY_TYPES = frozenset({"project", "epic", "task"})
-
-
-async def _add_mcp_entity(
-    *,
-    title: str,
-    content: str,
-    entity_type: str,
-    category: str | None,
-    languages: list[str] | None,
-    tags: list[str] | None,
-    related_to: list[str] | None,
-    metadata: dict[str, Any] | None,
-    project: str | None,
-    priority: str | None,
-    assignees: list[str] | None,
-    due_date: str | None,
-    technologies: list[str] | None,
-    depends_on: list[str] | None,
-    repository_url: str | None,
-    check_conflicts: bool = True,
-    skip_conflicts: bool = False,
-    conflict_threshold: float = 0.85,
-) -> dict[str, Any]:
-    from sibyl_core.tools.core import add
-
-    ctx = await _require_mcp_context(write=True)
-    normalized_entity_type = entity_type.strip().lower()
-    accessible_projects = await _resolve_mcp_project_scope(
-        ctx,
-        project,
-        require_project_when_restricted=True,
-    )
-    memory_scope = "project" if project else "private"
-    scope_key = project
-    write_decision = _authorize_mcp_memory_write(
-        ctx=ctx,
-        memory_scope=memory_scope,
-        scope_key=scope_key,
-        accessible_projects=accessible_projects,
-        surface="mcp_add",
-    )
-    await _validate_mcp_relationship_targets(
-        ctx=ctx,
-        related_to=related_to,
-        accessible_projects=accessible_projects,
-    )
-
-    authorized_scope = None if normalized_entity_type in _UNSCOPED_ENTITY_TYPES else memory_scope
-    full_metadata = stamp_memory_scope_metadata(
-        metadata,
-        memory_scope=authorized_scope,
-        scope_key=scope_key,
-        principal_id=ctx.user_id,
-    )
-    full_metadata["organization_id"] = ctx.org_id
-    if ctx.user_id:
-        full_metadata["created_by"] = ctx.user_id
-
-    add_kwargs = {
-        "title": title,
-        "content": content,
-        "entity_type": entity_type,
-        "category": category,
-        "languages": languages,
-        "tags": tags,
-        "related_to": related_to,
-        "metadata": full_metadata,
-        "project": project,
-        "priority": priority,
-        "assignees": assignees,
-        "due_date": due_date,
-        "technologies": technologies,
-        "depends_on": depends_on,
-        "repository_url": repository_url,
-        "check_conflicts": check_conflicts,
-        "skip_conflicts": skip_conflicts,
-        "conflict_threshold": max(conflict_threshold, 0.85),
-        "memory_scope": authorized_scope,
-        "scope_key": scope_key,
-        "principal_id": ctx.user_id,
-        # A declared supersedes may only name a row this caller can read, so
-        # the gate gets the same scope resolution the rest of the request used.
-        "accessible_projects": accessible_projects,
-        "allowed_memory_scope_keys": ctx.api_key_memory_scope_keys,
-    }
-    if normalized_entity_type == "project":
-        add_kwargs["sync"] = True
-
-    result = await add(**add_kwargs)
-    payload = _to_dict(result)
-    if normalized_entity_type == "project" and payload.get("success") and payload.get("id"):
-        if not ctx.user_id:
-            raise ValueError("user_context_required")
-        await create_project_record(
-            organization_id=UUID(ctx.org_id),
-            owner_user_id=UUID(ctx.user_id),
-            graph_project_id=str(payload["id"]),
-            name=title,
-            description=content,
-        )
-    payload["policy_reason"] = write_decision.reason
-    return payload
-
-
-async def _mcp_entity_write_target(
-    *, organization_id: str, entity_id: str
-) -> tuple[str | None, dict[str, Any]]:
-    """The project a manage target is addressed by, and its own scope metadata.
-
-    Resolving only the project made every entity-targeted action a project
-    write, so a contributor could mutate another principal's private row that
-    happened to live in a shared project. The scope travels with it so the
-    caller's authorization can answer to the row it is actually touching.
-    """
-    from sibyl_core.services.graph import get_surreal_graph_runtime
-    from sibyl_core.tools.helpers import _project_id_for_policy
-
-    runtime = await get_surreal_graph_runtime(organization_id)
-    entity = await runtime.entity_manager.get(entity_id)
-    if entity is None:
-        return None, {}
-    metadata = getattr(entity, "metadata", None)
-    return _project_id_for_policy(entity), dict(metadata) if isinstance(metadata, dict) else {}
-
-
-async def _authorize_mcp_manage_action(
-    *,
-    ctx: McpContext,
-    action: str,
-    entity_id: str | None,
-    accessible_projects: set[str] | None,
-) -> MemoryPolicyDecision | None:
-    normalized_action = action.lower().strip()
-    if normalized_action == "correct_memory":
-        if not entity_id:
-            return None
-        memory = await get_raw_memory(
-            organization_id=ctx.org_id,
-            memory_id=entity_id,
-        )
-        if memory is None:
-            memory = await get_raw_memory_by_source_id(
-                organization_id=ctx.org_id,
-                source_id=entity_id,
-            )
-        if memory is None:
-            return None
-        if memory.memory_scope is MemoryScope.PRIVATE and memory.principal_id != ctx.user_id:
-            raise ValueError("principal_mismatch")
-        policy_projects = accessible_projects
-        if policy_projects is None and memory.scope_key:
-            policy_projects = {memory.scope_key}
-        accessible_teams = (
-            await resolve_accessible_team_scope_keys(
-                user_id=ctx.user_id,
-                org_id=ctx.org_id,
-                scopes=ctx.scopes,
-            )
-            if memory.memory_scope is MemoryScope.TEAM and ctx.user_id
-            else None
-        )
-        return _authorize_mcp_memory_write(
-            ctx=ctx,
-            memory_scope=memory.memory_scope.value,
-            scope_key=memory.scope_key,
-            accessible_projects=policy_projects,
-            surface="mcp_manage_correct_memory",
-            accessible_teams=accessible_teams,
-        )
-    if normalized_action in MCP_PROJECT_ID_POLICY_ACTIONS:
-        project_id = entity_id
-    elif normalized_action in MCP_ENTITY_PROJECT_POLICY_ACTIONS:
-        if not entity_id:
-            return None
-        project_id, target_metadata = await _mcp_entity_write_target(
-            organization_id=ctx.org_id,
-            entity_id=entity_id,
-        )
-        target_scope = str(target_metadata.get("memory_scope") or "").strip()
-        if target_scope == MemoryScope.PRIVATE.value:
-            # A private row answers to its owner, never to project membership.
-            # The write policy authorizes a principal for their own private
-            # space, so the owner comparison has to happen here: the target is
-            # someone else's space, not the caller's.
-            owner = str(
-                target_metadata.get("principal_id") or target_metadata.get("scope_key") or ""
-            )
-            if not owner or not ctx.user_id or owner != str(ctx.user_id):
-                return MemoryPolicyDecision(
-                    action=MemoryPolicyAction.WRITE,
-                    allowed=False,
-                    reason="private_target_not_owned",
-                    memory_scope=MemoryScope.PRIVATE,
-                    scope_key=owner or None,
-                )
-            return _authorize_mcp_memory_write(
-                ctx=ctx,
-                memory_scope=MemoryScope.PRIVATE.value,
-                scope_key=None,
-                accessible_projects=accessible_projects,
-                surface="mcp_manage",
-            )
-    else:
-        return None
-
-    policy_projects = (
-        {project_id} if accessible_projects is None and project_id else accessible_projects
-    )
-    return _authorize_mcp_memory_write(
-        ctx=ctx,
-        memory_scope="project",
-        scope_key=project_id,
-        accessible_projects=policy_projects,
-        surface="mcp_manage",
-    )
-
-
-# MCP transition action -> the unified work-item action. Routing these through
-# the shared apps/api workflow service is what gives the MCP path the entity
-# lock, WebSocket broadcast, and project-activity bump that REST already had and
-# the old core manage() body silently skipped (audit H8).
-_MCP_WORKFLOW_TRANSITIONS: dict[str, WorkItemAction] = {
-    "start_task": WorkItemAction.START_TASK,
-    "block_task": WorkItemAction.BLOCK_TASK,
-    "unblock_task": WorkItemAction.UNBLOCK_TASK,
-    "submit_review": WorkItemAction.SUBMIT_REVIEW,
-    "complete_task": WorkItemAction.COMPLETE_TASK,
-    "archive_task": WorkItemAction.ARCHIVE_TASK,
-    "start_epic": WorkItemAction.START_EPIC,
-    "complete_epic": WorkItemAction.COMPLETE_EPIC,
-    "archive_epic": WorkItemAction.ARCHIVE_EPIC,
-}
-
-
-def _mcp_transition_message(action: str, *, learnings: str | None, reason: str | None) -> str:
-    """Mirror core manage()'s per-action success message for MCP clients."""
-    if action == "start_task":
-        return "Task started"
-    if action == "block_task":
-        return f"Task blocked: {reason or 'No reason provided'}"
-    if action == "unblock_task":
-        return "Task unblocked, resuming work"
-    if action == "submit_review":
-        return "Task submitted for review"
-    if action == "complete_task":
-        return "Task completed" + (" with learnings captured" if learnings else "")
-    if action == "archive_task":
-        return "Task archived"
-    if action == "start_epic":
-        return "Epic started"
-    if action == "complete_epic":
-        return "Epic completed" + (" with learnings captured" if learnings else "")
-    # archive_epic
-    return "Epic archived" + (f": {reason}" if reason else "")
-
-
-def _mcp_mutation_receipt(
-    data: Mapping[str, Any],
-    *,
-    applied: bool,
-    revision: int | None,
-    affected_records: list[str],
-) -> dict[str, Any]:
-    key = data.get("idempotency_key")
-    idempotency_key = key.strip() if isinstance(key, str) and key.strip() else None
-    return {
-        "operation_id": idempotency_key or str(uuid4()),
-        "applied": applied,
-        "revision": revision,
-        "affected_records": affected_records,
-        "idempotency_key": idempotency_key,
-        "replayed": False,
-    }
-
-
-async def _manage_workflow_transition(
-    *,
-    ctx: McpContext,
-    action: str,
-    work_item_action: WorkItemAction,
-    entity_id: str | None,
-    data: dict[str, Any],
-    policy_decision: MemoryPolicyDecision | None,
-) -> dict[str, Any]:
-    """Run an MCP task/epic transition through the shared workflow service.
-
-    The lock, broadcast, and project-activity bump come from the service by
-    construction. The response is shaped like core manage()'s ManageResponse so
-    MCP clients see no change, including the deprecation pointer.
-    """
-    from sibyl.locks import LockAcquisitionError
-    from sibyl.services.work_item_workflow import EPIC_TRANSITIONS, transition_work_item
-    from sibyl_core.errors import EntityNotFoundError, InvalidTransitionError, RevisionConflictError
-    from sibyl_core.tools.manage import _deprecation_notice
-
-    def _response(
-        *,
-        success: bool,
-        message: str,
-        action_data: dict[str, Any] | None = None,
-    ) -> dict[str, Any]:
-        payload: dict[str, Any] = {
-            "success": success,
-            "action": action,
-            "entity_id": entity_id,
-            "message": message,
-            "data": action_data or {},
-        }
-        deprecation = _deprecation_notice(action)
-        if deprecation is not None:
-            payload["data"].setdefault("deprecation", deprecation)
-        if policy_decision is not None:
-            payload["policy_reason"] = policy_decision.reason
-        return payload
-
-    if not entity_id:
-        suffix = "epic actions" if work_item_action in EPIC_TRANSITIONS else "task actions"
-        return _response(success=False, message=f"entity_id required for {suffix}")
-
-    entity: Any | None = None
-    if work_item_action in EPIC_TRANSITIONS:
-        from sibyl_core.models.entities import EntityType
-        from sibyl_core.services.graph import get_surreal_graph_runtime
-
-        runtime = await get_surreal_graph_runtime(ctx.org_id)
-        try:
-            entity = await runtime.entity_manager.get(entity_id)
-        except EntityNotFoundError:
-            entity = None
-        if not entity:
-            return _response(success=False, message=f"Epic not found: {entity_id}")
-        if entity.entity_type != EntityType.EPIC:
-            return _response(success=False, message=f"Entity is not an epic: {entity_id}")
-
-    learnings = data.get("learnings") if action in {"complete_task", "complete_epic"} else None
-    reason = data.get("reason") if action in {"block_task", "archive_epic"} else None
-    expected_revision = data.get("expected_revision")
-    if expected_revision is not None and (
-        isinstance(expected_revision, bool)
-        or not isinstance(expected_revision, int)
-        or expected_revision < 1
-    ):
-        return _response(
-            success=False,
-            message="expected_revision must be an integer greater than zero",
-        )
-
-    try:
-        transition_kwargs: dict[str, Any] = {"payload": data, "entity": entity}
-        if expected_revision is not None:
-            transition_kwargs["expected_revision"] = expected_revision
-        result = await transition_work_item(
-            ctx.org_id,
-            entity_id,
-            work_item_action,
-            **transition_kwargs,
-        )
-    except InvalidTransitionError as exc:
-        return _response(success=False, message=str(exc), action_data=exc.details)
-    except LockAcquisitionError:
-        return _response(
-            success=False,
-            message=f"{entity_id} is locked by another writer; retry shortly",
-        )
-    except (EntityNotFoundError, KeyError):
-        return _response(success=False, message=f"Work item not found: {entity_id}")
-    except RevisionConflictError as exc:
-        return _response(success=False, message=str(exc), action_data=exc.details)
-
-    # complete_task with learnings enqueues the same background learning jobs the
-    # REST surface does, using the policy context the authz step resolved.
-    response_data = dict(result.response_data)
-    response_data["mutation_receipt"] = _mcp_mutation_receipt(
-        data,
-        applied=True,
-        revision=result.revision,
-        affected_records=[f"entity:{entity_id}"],
-    )
-    cited_ids = data.get("cited_ids") if action == "complete_task" else None
-    if cited_ids:
-        from sibyl_core.tools.usage_citation import record_cited_item_usages
-
-        policy_context = policy_decision.policy_context if policy_decision else None
-        policy_project_id = None
-        if policy_context is not None:
-            policy_project_id = policy_context.project_id or (
-                policy_context.scope_key if policy_context.memory_space == "project" else None
-            )
-        try:
-            response_data["citation_usage"] = await record_cited_item_usages(
-                cited_ids,
-                organization_id=ctx.org_id,
-                principal_id=ctx.user_id,
-                project_id=result.task_data.get("project_id") or policy_project_id,
-                source_surface="mcp_manage_complete_task",
-                request_metadata={
-                    "action": action,
-                    "has_learnings": bool(learnings),
-                    "task_id": entity_id,
-                },
-            )
-        except Exception as exc:
-            log.warning(
-                "mcp_complete_task_citation_usage_failed",
-                task_id=entity_id,
-                error_type=type(exc).__name__,
-                exc_info=True,
-            )
-    if (
-        action == "complete_task"
-        and learnings
-        and policy_decision is not None
-        and policy_decision.policy_context is not None
-    ):
-        from sibyl.jobs.entities import serialize_memory_policy_context
-        from sibyl.jobs.queue import (
-            enqueue_create_learning_episode,
-            enqueue_create_learning_procedure,
-        )
-
-        policy_payload = serialize_memory_policy_context(policy_decision.policy_context)
-        episode_job_id = await enqueue_create_learning_episode(
-            result.task_data, ctx.org_id, policy_context=policy_payload
-        )
-        procedure_job_id = await enqueue_create_learning_procedure(
-            result.task_data, ctx.org_id, policy_context=policy_payload
-        )
-        response_data["learning_episode_job_id"] = episode_job_id
-        response_data["learning_procedure_job_id"] = procedure_job_id
-
-    return _response(
-        success=True,
-        message=_mcp_transition_message(action, learnings=learnings, reason=reason),
-        action_data=response_data,
-    )
-
-
-async def _manage_memory_correction(
-    *,
-    ctx: McpContext,
-    entity_id: str | None,
-    data: dict[str, Any],
-    accessible_projects: set[str] | None,
-    policy_decision: MemoryPolicyDecision | None,
-) -> dict[str, Any]:
-    from sibyl_core.services.memory import apply_memory_correction
-
-    if not entity_id:
-        return {
-            "success": False,
-            "action": "correct_memory",
-            "entity_id": None,
-            "message": "entity_id (raw memory ID) required for correct_memory action",
-            "data": {},
-        }
-    correction_action = data.get("action")
-    if not isinstance(correction_action, str) or not correction_action.strip():
-        return {
-            "success": False,
-            "action": "correct_memory",
-            "entity_id": entity_id,
-            "message": "data.action required for correct_memory action",
-            "data": {},
-        }
-    reason = data.get("reason")
-    if not isinstance(reason, str) or not reason.strip():
-        return {
-            "success": False,
-            "action": "correct_memory",
-            "entity_id": entity_id,
-            "message": "data.reason required for correct_memory action",
-            "data": {},
-        }
-    expected_revision = data.get("expected_revision")
-    if expected_revision is not None and (
-        isinstance(expected_revision, bool)
-        or not isinstance(expected_revision, int)
-        or expected_revision < 1
-    ):
-        return {
-            "success": False,
-            "action": "correct_memory",
-            "entity_id": entity_id,
-            "message": "expected_revision must be an integer greater than zero",
-            "data": {},
-        }
-    memory = await get_raw_memory(organization_id=ctx.org_id, memory_id=entity_id)
-    if memory is None:
-        memory = await get_raw_memory_by_source_id(
-            organization_id=ctx.org_id,
-            source_id=entity_id,
-        )
-    if memory is None:
-        return {
-            "success": False,
-            "action": "correct_memory",
-            "entity_id": entity_id,
-            "message": f"Memory source not found: {entity_id}",
-            "data": {},
-        }
-    result = await apply_memory_correction(
-        organization_id=ctx.org_id,
-        source_id=memory.id,
-        principal_id=ctx.user_id,
-        action=correction_action,
-        reason=reason.strip(),
-        accessible_projects=(
-            {memory.scope_key}
-            if accessible_projects is None
-            and memory.memory_scope is MemoryScope.PROJECT
-            and memory.scope_key
-            else accessible_projects
-        ),
-        accessible_teams=(
-            policy_decision.policy_context.accessible_teams
-            if policy_decision is not None and policy_decision.policy_context is not None
-            else None
-        ),
-        replacement_source_id=(
-            str(data["replacement_source_id"]) if data.get("replacement_source_id") else None
-        ),
-        duplicate_of_source_id=(
-            str(data["duplicate_of_source_id"]) if data.get("duplicate_of_source_id") else None
-        ),
-        revised_content=(
-            str(data["revised_content"]) if data.get("revised_content") is not None else None
-        ),
-        expected_revision=expected_revision,
-    )
-    revision = result.updated_memory.revision if result.updated_memory else None
-    affected_records = (
-        [f"raw_captures:{source_id}" for source_id in result.preview.affected_source_ids]
-        if result.applied
-        else []
-    )
-    response_data = {
-        "allowed": result.preview.allowed,
-        "applied": result.applied,
-        "correction_action": result.preview.action,
-        "reason": result.preview.reason,
-        "revision": revision,
-        "affected_source_ids": list(result.preview.affected_source_ids),
-        "affected_derived_ids": list(result.preview.affected_derived_ids),
-        "mutation_receipt": _mcp_mutation_receipt(
-            data,
-            applied=result.applied,
-            revision=revision,
-            affected_records=affected_records,
-        ),
-    }
-    await log_memory_audit_event(
-        action=result.preview.audit_action,
-        user_id=ctx.user_id,
-        organization_id=ctx.org_id,
-        request=None,
-        memory_scope=memory.memory_scope.value,
-        scope_key=memory.scope_key,
-        project_id=memory.project_id or memory.scope_key,
-        source_surface="mcp_manage_correct_memory",
-        source_ids=result.preview.affected_source_ids or [memory.id],
-        derived_ids=result.preview.affected_derived_ids,
-        policy_allowed=result.preview.allowed and result.applied,
-        policy_reason=result.preview.reason,
-        details={
-            "action": result.preview.action,
-            "applied": result.applied,
-            "target_lifecycle_state": result.preview.target_lifecycle_state,
-            "target_lifecycle_flags": result.preview.target_lifecycle_flags,
-        },
-    )
-    payload = {
-        "success": result.applied,
-        "action": "correct_memory",
-        "entity_id": entity_id,
-        "message": (
-            f"Memory correction applied: {result.preview.action}"
-            if result.applied
-            else f"Memory correction denied: {result.preview.reason}"
-        ),
-        "data": response_data,
-    }
-    if policy_decision is not None:
-        payload["policy_reason"] = policy_decision.reason
-    return payload
-
-
-@_serialize_mcp_idempotency("mcp/manage", action_scoped=True)
-async def _manage_mcp_action(
-    *,
-    action: str,
-    entity_id: str | None,
-    data: dict[str, Any] | None,
-) -> dict[str, Any]:
-    from sibyl_core.tools.manage import manage
-
-    ctx = await _require_mcp_context(write=True)
-    accessible_projects = await _get_accessible_projects(ctx)
-    policy_decision = await _authorize_mcp_manage_action(
-        ctx=ctx,
-        action=action,
-        entity_id=entity_id,
-        accessible_projects=accessible_projects,
-    )
-    # A denied decision has to stop the call, not annotate it. This ran before
-    # the idempotency reservation deliberately: a refused action must not
-    # strand a reservation row that a later legitimate retry would replay.
-    if policy_decision is not None and not policy_decision.allowed:
-        raise ValueError(policy_decision.reason)
-
-    normalized_action = action.lower().strip()
-    request_data = dict(data or {})
-    idempotency_key = request_data.get("idempotency_key")
-    if idempotency_key is not None and (
-        not isinstance(idempotency_key, str)
-        or not idempotency_key.strip()
-        or len(idempotency_key.strip()) > 255
-    ):
-        return {
-            "success": False,
-            "action": normalized_action,
-            "entity_id": entity_id,
-            "message": "idempotency_key must be a non-empty string up to 255 characters",
-            "data": {},
-        }
-    idempotency_key = idempotency_key.strip() if isinstance(idempotency_key, str) else None
-    if idempotency_key is not None:
-        request_data["idempotency_key"] = idempotency_key
-    idempotency_path = f"mcp/manage/{normalized_action}"
-    idempotency_payload = {"entity_id": entity_id, "data": dict(request_data)}
-    principal_id = ctx.user_id or "unknown"
-    idempotency_claim: ApiIdempotencyRecord | None = None
-    if idempotency_key is not None:
-        record, claimed = await reserve_idempotency_record(
-            organization_id=UUID(ctx.org_id),
-            principal_id=principal_id,
-            idempotency_key=idempotency_key,
-            method="MCP",
-            path=idempotency_path,
-            payload=idempotency_payload,
-            content_session=None,
-        )
-        if not claimed:
-            if record.request_hash != idempotency_request_hash(idempotency_payload):
-                return {
-                    "success": False,
-                    "action": normalized_action,
-                    "entity_id": entity_id,
-                    "message": "idempotency_key was already used for a different request",
-                    "data": {},
-                }
-            if idempotency_record_pending(record):
-                # This branch runs under _serialize_mcp_idempotency's lock, so
-                # no executor is live on this key: the reservation was
-                # interrupted before its receipt completed. Adopt it and
-                # re-execute; completion overwrites the same record id.
-                log.warning(
-                    "mcp_idempotency_interrupted_takeover",
-                    path=idempotency_path,
-                    organization_id=ctx.org_id,
-                )
-                idempotency_claim = record
-            else:
-                replayed = cast("dict[str, Any]", deepcopy(record.response_body))
-                raw_response_data = replayed.get("data")
-                if isinstance(raw_response_data, dict):
-                    response_data = cast("dict[str, Any]", raw_response_data)
-                    receipt = response_data.get("mutation_receipt")
-                    if isinstance(receipt, dict):
-                        response_data["mutation_receipt"] = {**receipt, "replayed": True}
-                return replayed
-        else:
-            idempotency_claim = record
-
-    full_data = request_data
-    full_data["organization_id"] = ctx.org_id
-    if ctx.user_id:
-        full_data["user_id"] = ctx.user_id
-
-    # Task/epic transitions route through the shared workflow service so the MCP
-    # path gains locking, broadcasting, and project-activity by construction.
-    # Everything else (update_task, add_note, crawl, analysis, ...) stays on the
-    # core manage() dispatcher unchanged.
-    work_item_action = _MCP_WORKFLOW_TRANSITIONS.get(normalized_action)
-    if work_item_action is not None:
-        payload = await _manage_workflow_transition(
-            ctx=ctx,
-            action=normalized_action,
-            work_item_action=work_item_action,
-            entity_id=entity_id,
-            data=full_data,
-            policy_decision=policy_decision,
-        )
-    elif normalized_action == "correct_memory":
-        payload = await _manage_memory_correction(
-            ctx=ctx,
-            entity_id=entity_id,
-            data=full_data,
-            accessible_projects=accessible_projects,
-            policy_decision=policy_decision,
-        )
-    else:
-        result = await manage(
-            action=action,
-            entity_id=entity_id,
-            data=full_data,
-            organization_id=ctx.org_id,
-            principal_id=ctx.user_id,
-            accessible_projects=accessible_projects,
-            allowed_memory_scope_keys=ctx.api_key_memory_scope_keys,
-        )
-        payload = _to_dict(result)
-        if policy_decision is not None:
-            payload["policy_reason"] = policy_decision.reason
-
-    if idempotency_claim is not None:
-        await complete_idempotency_record(
-            idempotency_claim,
-            response_status_code=200,
-            response_body=payload,
-            content_session=None,
-        )
-    return payload
-
-
-async def _require_org_id() -> str:
-    """Require organization ID from MCP context.
-
-    Raises:
-        ValueError: If no organization context is available.
-
-    Returns:
-        The organization ID string.
-    """
-    ctx = await _require_mcp_context()
-    return ctx.org_id
-
-
-async def _require_owner_mcp_context(ctx: McpContext) -> None:
-    """Require OWNER membership for the current MCP context."""
-    if not await has_owner_membership(org_id=ctx.org_id, user_id=ctx.user_id):
-        raise ValueError("OWNER role required for log access")
+from sibyl.mcp_tools.registration import register_tools
 
 
 def create_mcp_server() -> MCPServer:
@@ -1963,946 +67,9 @@ def create_mcp_server() -> MCPServer:
         async def _oauth_org_post(request):
             return await auth_server_provider.ui_org_post(request)
 
-    _register_tools(mcp)
+    register_tools(mcp)
     _register_resources(mcp)
     return mcp
-
-
-def _to_dict(obj: Any) -> Any:
-    """Convert dataclass or object to dict for JSON serialization."""
-    if hasattr(obj, "__dataclass_fields__"):
-        return asdict(obj)
-    if isinstance(obj, list):
-        return [_to_dict(item) for item in obj]
-    return obj
-
-
-def _register_tools(mcp: MCPServer) -> None:
-    """Register all MCP tools on the server instance."""
-
-    # =========================================================================
-    # TOOL 1: search - UNIFIED SEARCH (graph + documents)
-    # =========================================================================
-
-    @mcp.tool()
-    async def search(
-        query: str,
-        types: list[str] | None = None,
-        language: str | None = None,
-        category: str | None = None,
-        status: str | None = None,
-        project: str | None = None,
-        source: str | None = None,
-        source_id: str | None = None,
-        source_name: str | None = None,
-        assignee: str | None = None,
-        since: str | None = None,
-        limit: int = 10,
-        include_content: bool = True,
-        content_max_chars: int = 500,
-        include_documents: bool = True,
-        include_graph: bool = True,
-        use_enhanced: bool = True,
-        boost_recent: bool = True,
-        temporal_decay_days: float | None = None,
-    ) -> dict[str, Any]:
-        """Unified semantic search across knowledge graph AND documentation.
-
-        Searches both Sibyl's knowledge graph (patterns, rules, episodes, tasks)
-        AND crawled documentation (Surreal-backed vector search). Results are
-        merged and ranked by relevance score.
-
-        IMPORTANT FOR AGENTS:
-        - Results contain PREVIEWS only (truncated content)
-        - To get FULL content, use: sibyl show <id>
-        - Do NOT try to read URLs directly - content is stored in Sibyl
-        - The 'id' field is the entity/chunk ID to fetch full content
-
-        Args:
-            query: Natural language search query
-            types: Entity types to search. Options: pattern, rule, template,
-                   topic, episode, task, project, document.
-                   Include 'document' to search crawled docs.
-            language: Filter by programming language (python, typescript, etc.)
-            category: Filter by category/domain (authentication, database, etc.)
-            status: Filter tasks by status (backlog, todo, doing, blocked, review, done)
-            project: Filter tasks by project ID
-            source: Alias for source_name (for convenience)
-            source_id: Filter documents by source UUID
-            source_name: Filter documents by source name (partial match)
-            assignee: Filter tasks by assignee name
-            since: Filter by creation date (ISO format: 2024-03-15 or relative: 7d, 2w)
-            limit: Maximum results to return (1-50, default: 10)
-            include_content: Include content in results (default: True)
-            content_max_chars: Maximum content characters per result (default: 500)
-            include_documents: Search crawled documentation (default: True)
-            include_graph: Search knowledge graph entities (default: True)
-            use_enhanced: Use enhanced hybrid retrieval, vector + graph fusion (default: True)
-            boost_recent: Boost recent results in ranking (default: True)
-
-        Returns:
-            Search results with:
-            - id: Entity/chunk ID (use with 'sibyl show <id>' for full content)
-            - type: Entity type (pattern, rule, task, document, etc.)
-            - name: Title/name of the result
-            - content: PREVIEW only - truncated, use show for full content
-            - score: Relevance score (0-1)
-            - source: Source name for documentation results
-            - result_origin: "graph" or "document" indicating data source
-            - usage_hint: Instructions for getting full content
-
-        Examples:
-            # Search everything
-            search("authentication patterns")
-
-            # Search only documentation
-            search("Next.js middleware", include_graph=False)
-
-            # Get full content of a result
-            # 1. search("OAuth") -> returns results with IDs
-            # 2. sibyl show <id> -> returns full content
-        """
-        from sibyl_core.tools.core import search as _search
-
-        # Get full context from authenticated MCP session
-        ctx = await _require_mcp_context()
-        accessible_projects = await _get_accessible_projects(ctx)
-        api_key_memory_scope_keys = ctx.api_key_memory_scope_keys
-
-        result = await _search(
-            query=query,
-            types=types,
-            language=language,
-            category=category,
-            status=status,
-            project=project,
-            accessible_projects=accessible_projects,
-            source=source,
-            source_id=source_id,
-            source_name=source_name,
-            assignee=assignee,
-            since=since,
-            limit=limit,
-            include_content=include_content,
-            content_max_chars=content_max_chars,
-            include_documents=include_documents,
-            include_graph=include_graph,
-            use_enhanced=use_enhanced,
-            boost_recent=boost_recent,
-            temporal_decay_days=temporal_decay_days,
-            organization_id=ctx.org_id,
-            principal_id=getattr(ctx, "user_id", None),
-            allowed_memory_scope_keys=(
-                set(api_key_memory_scope_keys) if api_key_memory_scope_keys is not None else None
-            ),
-        )
-        return _to_dict(result)
-
-    # =========================================================================
-    # TOOL 2: context
-    # =========================================================================
-
-    @mcp.tool()
-    async def context(
-        goal: str,
-        intent: Literal[
-            "build", "plan", "ideate", "research", "debug", "decide", "learn", "general"
-        ] = "build",
-        layer: Literal["wake", "recall", "deep_search"] = "recall",
-        domain: str | None = None,
-        project: str | None = None,
-        agent_id: str | None = None,
-        limit: int = 24,
-        include_related: bool = True,
-        related_limit: int = 3,
-        audit: bool = False,
-        markdown_token_budget: int | None = DEFAULT_MARKDOWN_TOKEN_BUDGET,
-    ) -> dict[str, Any]:
-        """Compile a precise context pack for an agent goal.
-
-        Context packs are structured for action, not generic search browsing.
-        They group relevant memories into facets like active work, decisions,
-        plans, ideas, constraints, artifacts, procedures, gotchas, and recent
-        sessions. Use this before dispatching or resuming agents.
-
-        Args:
-            goal: What the agent is trying to accomplish.
-            intent: Goal mode - build, plan, ideate, research, debug, decide,
-                learn, or general.
-            layer: Retrieval depth - wake for compact session start, recall for
-                working context, or deep_search for broad research.
-            domain: Optional domain/category to scope context. This can be
-                software, creative work, home projects, research, or any other
-                modeled domain.
-            project: Optional project ID to scope active work.
-            agent_id: Optional agent diary identity to include alongside normal
-                private/project raw memory.
-            limit: Maximum total context items, clamped to 1-50.
-            include_related: Include one-hop related graph context.
-            related_limit: Related items per selected context item.
-            audit: Include full retrieval metadata per item for pack auditing.
-            markdown_token_budget: Cap rendered markdown at roughly this many
-                tokens for small-context consumers.
-        """
-        return await _compile_mcp_context_pack(
-            goal=goal,
-            intent=intent,
-            layer=layer,
-            domain=domain,
-            project=project,
-            agent_id=agent_id,
-            limit=limit,
-            include_related=include_related,
-            related_limit=related_limit,
-            audit=audit,
-            markdown_token_budget=markdown_token_budget,
-        )
-
-    # =========================================================================
-    # TOOL 3: synthesis_plan
-    # =========================================================================
-
-    @mcp.tool()
-    async def synthesis_plan(
-        goal: str,
-        output_type: SynthesisOutputKind = "documentation",
-        audience: str | None = None,
-        depth: SynthesisDepthKind = "standard",
-        seed_query: str | None = None,
-        project: str | None = None,
-        domain: str | None = None,
-        entity_ids: list[str] | None = None,
-        decision_ids: list[str] | None = None,
-        task_ids: list[str] | None = None,
-        artifact_ids: list[str] | None = None,
-        required_sections: list[dict[str, Any] | str] | None = None,
-        constraints: list[str] | None = None,
-        max_sections: int = 6,
-        include_neighborhoods: bool = True,
-    ) -> dict[str, Any]:
-        """Plan source-grounded synthesis from authorized memory."""
-        return await _synthesis_mcp_plan(
-            goal=goal,
-            output_type=output_type,
-            audience=audience,
-            depth=depth,
-            seed_query=seed_query,
-            project=project,
-            domain=domain,
-            entity_ids=entity_ids,
-            decision_ids=decision_ids,
-            task_ids=task_ids,
-            artifact_ids=artifact_ids,
-            required_sections=required_sections,
-            constraints=constraints,
-            max_sections=max_sections,
-            include_neighborhoods=include_neighborhoods,
-        )
-
-    # =========================================================================
-    # TOOL 4: synthesis_draft
-    # =========================================================================
-
-    @mcp.tool()
-    async def synthesis_draft(
-        goal: str,
-        output_type: SynthesisOutputKind = "documentation",
-        audience: str | None = None,
-        depth: SynthesisDepthKind = "standard",
-        seed_query: str | None = None,
-        project: str | None = None,
-        domain: str | None = None,
-        entity_ids: list[str] | None = None,
-        decision_ids: list[str] | None = None,
-        task_ids: list[str] | None = None,
-        artifact_ids: list[str] | None = None,
-        required_sections: list[dict[str, Any] | str] | None = None,
-        constraints: list[str] | None = None,
-        max_sections: int = 6,
-        include_neighborhoods: bool = True,
-        output_format: SynthesisArtifactKind = "markdown",
-        remember: bool = False,
-        memory_scope: str = "private",
-        scope_key: str | None = None,
-        tags: list[str] | None = None,
-    ) -> dict[str, Any]:
-        """Draft, verify, and optionally remember a source-grounded artifact."""
-        return await _synthesis_mcp_draft(
-            goal=goal,
-            output_type=output_type,
-            audience=audience,
-            depth=depth,
-            seed_query=seed_query,
-            project=project,
-            domain=domain,
-            entity_ids=entity_ids,
-            decision_ids=decision_ids,
-            task_ids=task_ids,
-            artifact_ids=artifact_ids,
-            required_sections=required_sections,
-            constraints=constraints,
-            max_sections=max_sections,
-            include_neighborhoods=include_neighborhoods,
-            output_format=output_format,
-            remember=remember,
-            memory_scope=memory_scope,
-            scope_key=scope_key,
-            tags=tags,
-        )
-
-    # =========================================================================
-    # TOOL 5: synthesis_verify
-    # =========================================================================
-
-    @mcp.tool()
-    async def synthesis_verify(
-        goal: str,
-        output_type: SynthesisOutputKind = "documentation",
-        audience: str | None = None,
-        depth: SynthesisDepthKind = "standard",
-        seed_query: str | None = None,
-        project: str | None = None,
-        domain: str | None = None,
-        entity_ids: list[str] | None = None,
-        decision_ids: list[str] | None = None,
-        task_ids: list[str] | None = None,
-        artifact_ids: list[str] | None = None,
-        required_sections: list[dict[str, Any] | str] | None = None,
-        constraints: list[str] | None = None,
-        max_sections: int = 6,
-        include_neighborhoods: bool = True,
-    ) -> dict[str, Any]:
-        """Verify citation, hidden-context, freshness, and gap coverage."""
-        return await _synthesis_mcp_verify(
-            goal=goal,
-            output_type=output_type,
-            audience=audience,
-            depth=depth,
-            seed_query=seed_query,
-            project=project,
-            domain=domain,
-            entity_ids=entity_ids,
-            decision_ids=decision_ids,
-            task_ids=task_ids,
-            artifact_ids=artifact_ids,
-            required_sections=required_sections,
-            constraints=constraints,
-            max_sections=max_sections,
-            include_neighborhoods=include_neighborhoods,
-        )
-
-    # =========================================================================
-    # TOOL 6: explore
-    # =========================================================================
-
-    @mcp.tool()
-    async def explore(
-        mode: Literal["list", "related", "traverse", "dependencies"] = "list",
-        types: list[str] | None = None,
-        entity_id: str | None = None,
-        relationship_types: list[str] | None = None,
-        depth: int = 1,
-        language: str | None = None,
-        category: str | None = None,
-        project: str | None = None,
-        status: str | None = None,
-        limit: int = 50,
-    ) -> dict[str, Any]:
-        """Explore and browse the knowledge graph.
-
-        Four modes of exploration:
-        - list: Browse entities by type with optional filters
-        - related: Find entities directly connected to a specific entity
-        - traverse: Multi-hop graph traversal from an entity
-        - dependencies: Task dependency chains in topological order
-
-        Args:
-            mode: Exploration mode - "list", "related", "traverse", or "dependencies"
-            types: Entity types to explore (for list mode)
-            entity_id: Starting entity ID (required for related/traverse/dependencies modes)
-            relationship_types: Filter by relationship types
-                               (APPLIES_TO, REQUIRES, CONFLICTS_WITH, SUPERSEDES,
-                                DOCUMENTED_IN, ENABLES, BREAKS, PART_OF, RELATED_TO,
-                                DERIVED_FROM)
-            depth: Traversal depth for traverse mode (1-3, default: 1)
-            language: Filter by programming language
-            category: Filter by category
-            project: Filter tasks by project ID (for list mode with tasks)
-            status: Filter tasks by status (for list mode with tasks)
-            limit: Maximum results (1-200, default: 50)
-
-        Returns:
-            Exploration results with entities and/or relationships
-
-        Examples:
-            explore(mode="list", types=["pattern"], language="typescript")
-            explore(mode="list", types=["task"], project="proj_abc", status="todo")
-            explore(mode="related", entity_id="pattern:error-handling")
-            explore(mode="traverse", entity_id="topic:auth", depth=2)
-            explore(mode="dependencies", entity_id="task_xyz")
-        """
-        from sibyl_core.tools.core import explore as _explore
-
-        # Get full context from authenticated MCP session
-        ctx = await _require_mcp_context()
-        accessible_projects = await _get_accessible_projects(ctx)
-
-        api_key_memory_scope_keys = ctx.api_key_memory_scope_keys
-        result = await _explore(
-            mode=mode,
-            types=types,
-            entity_id=entity_id,
-            relationship_types=relationship_types,
-            depth=depth,
-            language=language,
-            category=category,
-            project=project,
-            accessible_projects=accessible_projects,
-            status=status,
-            limit=limit,
-            organization_id=ctx.org_id,
-            principal_id=getattr(ctx, "user_id", None),
-            allowed_memory_scope_keys=(
-                set(api_key_memory_scope_keys) if api_key_memory_scope_keys is not None else None
-            ),
-        )
-        return _to_dict(result)
-
-    # =========================================================================
-    # TOOL 7: expand_neighbors
-    # =========================================================================
-
-    @mcp.tool()
-    async def expand_neighbors(
-        entity_ids: list[str],
-        relationship_types: list[str] | None = None,
-        types: list[str] | None = None,
-        depth: int = DEFAULT_TRAVERSAL_DEPTH,
-        limit: int = DEFAULT_EXPAND_LIMIT,
-        content_max_chars: int = DEFAULT_NEIGHBOR_CONTENT_MAX_CHARS,
-        include_incoming: bool = True,
-        project: str | None = None,
-    ) -> dict[str, Any]:
-        """Widen memories you already found into their graph neighborhood.
-
-        TRAVERSAL CONTRACT (read this before calling):
-        This is a bounded step in a loop of AT MOST THREE ROUNDS. Round one is
-        `search` or `context`. Round two widens with `expand_neighbors` or
-        `fetch_slice`. Round three widens once more, and then you answer.
-        Two widening rounds capture nearly all of the available gain; a fourth
-        costs latency and buys noise.
-
-        SKIP THIS VERB when one hop answers the question. If you want "what do we
-        know about X", call `context` and read the pack it composed. Traversal is
-        for when you have specific memories in hand and need what sits next to
-        them: the tasks blocking this one, the decision a plan supersedes, the
-        spans of a memory a search only matched part of.
-
-        COMPOSITION IS NOT YOURS. This verb returns previews and adjacency so you
-        can choose what to gather. `context` still renders the evidence, and the
-        reserved note lane and evidence ordering stay under its control.
-
-        Every neighbor is authorized for you individually. A neighbor missing
-        from the result may exist and be someone else's private memory; that is
-        the system working, not a gap to route around.
-
-        Args:
-            entity_ids: Seed entity IDs, at most 8 of them.
-                Seeds that resolve to nothing you may read come back in
-                `unresolved` without saying which reason applied.
-            relationship_types: Restrict hops to these relationship names, e.g.
-                ["DEPENDS_ON"] or ["PART_OF"]. Empty walks every relationship.
-            types: Restrict neighbors to these entity types.
-            depth: Hops to walk, 1-3 (default 1). Depth 1 first; deepen only
-                when depth 1 came back thin.
-            limit: Neighbors returned, up to 24 (default 8).
-            content_max_chars: Preview characters per neighbor. These are
-                previews by design; widen a promising one with `fetch_slice`.
-            include_incoming: Follow edges pointing at the seeds too (default
-                True). Dependents and passages are only reachable inbound.
-            project: Scope the walk to one project you can read.
-
-        Returns:
-            Hop-tagged neighbors with relationship, direction, and distance,
-            highest path score first, plus `truncated` when more existed.
-
-        Examples:
-            expand_neighbors(["task_abc"], relationship_types=["DEPENDS_ON"])
-            expand_neighbors(["decision_1", "decision_2"], depth=2)
-        """
-        from sibyl_core.tools.core import expand_neighbors as _expand_neighbors
-
-        ctx = await _require_mcp_context()
-        accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-        api_key_memory_scope_keys = ctx.api_key_memory_scope_keys
-        result = await _expand_neighbors(
-            entity_ids,
-            organization_id=ctx.org_id,
-            relationship_types=relationship_types,
-            types=types,
-            depth=depth,
-            limit=limit,
-            content_max_chars=content_max_chars,
-            include_incoming=include_incoming,
-            principal_id=getattr(ctx, "user_id", None),
-            accessible_projects=accessible_projects,
-            allowed_memory_scope_keys=(
-                set(api_key_memory_scope_keys) if api_key_memory_scope_keys is not None else None
-            ),
-        )
-        return _to_dict(result)
-
-    # =========================================================================
-    # TOOL 8: fetch_slice
-    # =========================================================================
-
-    @mcp.tool()
-    async def fetch_slice(
-        entity_id: str,
-        window: int = DEFAULT_SLICE_WINDOW,
-        content_max_chars: int = DEFAULT_SLICE_CONTENT_MAX_CHARS,
-        project: str | None = None,
-    ) -> dict[str, Any]:
-        """Read one memory at span granularity, centered where you point.
-
-        TRAVERSAL CONTRACT (read this before calling):
-        This is a bounded step in a loop of AT MOST THREE ROUNDS, the same budget
-        `expand_neighbors` spends from. Use it when a search hit is a passage, or
-        when a memory is long and you need the part around a match rather than
-        the whole body.
-
-        SKIP THIS VERB when the memory you found is already short enough to read.
-        A result whose content was not truncated needs no widening.
-
-        COMPOSITION IS NOT YOURS. This verb hands back spans, not an answer.
-        `context` composes the final evidence, and it keeps control of ordering
-        and the reserved note lane whatever you gather here.
-
-        CITE THE PARENT, NOT THE SPAN. The response names `parent_id`, and that
-        is the id a later reader can resolve. Span ids are re-minted whenever the
-        memory is edited, so a citation pointing at one goes stale silently.
-
-        A memory short enough never to have been cut comes back whole with
-        `sliced=false`. That is the answer, not an error to retry.
-
-        Args:
-            entity_id: A passage entity ID, or the ID of the memory it came from.
-                Given a passage, the window is centered on it. Given a memory,
-                the window starts at its first span.
-            window: Adjacent spans to return, 1-64. The default of
-                three is the measured adjacency: three spans reach the same
-                exposure as the whole memory, one span reaches noticeably less.
-            content_max_chars: Character budget for the whole window, spent in
-                span order. The span that exhausts it says `truncated`.
-            project: Scope the read to one project you can read.
-
-        Returns:
-            The ordered span window with per-span index and total, plus the
-            parent memory a citation resolves to.
-
-        Examples:
-            fetch_slice("passage_9f2c1b")
-            fetch_slice("decision_abc", window=5)
-        """
-        from sibyl_core.tools.core import fetch_slice as _fetch_slice
-
-        ctx = await _require_mcp_context()
-        accessible_projects = await _resolve_mcp_project_scope(ctx, project)
-        api_key_memory_scope_keys = ctx.api_key_memory_scope_keys
-        result = await _fetch_slice(
-            entity_id,
-            organization_id=ctx.org_id,
-            window=window,
-            content_max_chars=content_max_chars,
-            principal_id=getattr(ctx, "user_id", None),
-            accessible_projects=accessible_projects,
-            allowed_memory_scope_keys=(
-                set(api_key_memory_scope_keys) if api_key_memory_scope_keys is not None else None
-            ),
-        )
-        return _to_dict(result)
-
-    # =========================================================================
-    # TOOL 9: add
-    # =========================================================================
-
-    @mcp.tool()
-    async def add(
-        title: str,
-        content: str,
-        entity_type: str = "episode",
-        category: str | None = None,
-        languages: list[str] | None = None,
-        tags: list[str] | None = None,
-        related_to: DeclaredRelatedTo = None,
-        metadata: dict[str, Any] | None = None,
-        # Task-specific parameters
-        project: str | None = None,
-        priority: str | None = None,
-        assignees: list[str] | None = None,
-        due_date: str | None = None,
-        technologies: list[str] | None = None,
-        depends_on: list[str] | None = None,
-        # Project-specific parameters
-        repository_url: str | None = None,
-        # Conflict detection
-        check_conflicts: bool = True,
-        skip_conflicts: bool = False,
-        conflict_threshold: float = 0.85,
-    ) -> dict[str, Any]:
-        """Add new knowledge to the graph.
-
-        Creates a new knowledge entity that can be searched and explored.
-        Supports episodes, patterns, procedures, tasks, epics, projects, and
-        domain-general memories such as decisions, plans, ideas, claims,
-        artifacts, sessions, and domains.
-
-        ENTITY TYPES:
-        - episode: Temporal knowledge (default) - insights, learnings, discoveries
-        - pattern: Coding pattern or best practice
-        - procedure: Repeatable workflow or runbook
-        - decision: Chosen direction with rationale
-        - plan: Strategy, sequencing, milestones, or project plan
-        - idea: Brainstormed concept or unresolved option
-        - claim: Atomic fact or assertion with provenance/confidence
-        - artifact: File, object, document, asset, system, or work product
-        - session: Conversation or work-session checkpoint
-        - domain: Any modeled problem space, software or otherwise
-        - task: Work item with workflow state machine (REQUIRES project)
-        - epic: Feature initiative grouping tasks (REQUIRES project)
-        - project: Container for related tasks
-
-        Args:
-            title: Short title for the knowledge (max 200 chars)
-            content: Full content/description (max 50000 chars)
-            entity_type: Type such as episode, decision, plan, idea, claim,
-                artifact, procedure, task, epic, or project
-            category: Category for organization (e.g., "debugging", "architecture")
-            languages: Applicable programming languages
-            tags: Browse-only metadata tags for filtering and organization.
-                Tags do not affect ranked recall.
-            related_to: IDs of related entities to link. A bare ID links
-                untyped; prefixing an ID with a predicate declares what this
-                memory does to that one and retrieval weights the declared
-                predicate: "supersedes:<id>", "contradicts:<id>",
-                "requires:<id>", "supports:<id>", "decides:<id>". The memory
-                being written is always the subject.
-            metadata: Additional structured metadata (stored as JSON)
-            project: Project ID (REQUIRED for tasks). Use explore(types=["project"]) to find projects.
-            priority: Task priority - critical, high, medium (default), low, someday
-            assignees: List of assignee names for tasks
-            due_date: Due date for tasks (ISO format: 2024-03-15)
-            technologies: Technologies involved (for tasks)
-            depends_on: Task IDs this depends on (creates DEPENDS_ON edges)
-            repository_url: Repository URL for projects
-            check_conflicts: Check for semantically similar existing knowledge
-            skip_conflicts: Skip conflict detection for latency-sensitive captures
-            conflict_threshold: Similarity score required to flag a conflict
-
-        Returns:
-            Result with success status, entity ID, and message
-
-        Examples:
-            # Record a learning
-            add("Debug: Redis timeout", "Problem was connection pool exhaustion",
-                entity_type="pattern", category="debugging")
-
-            # Create a task (project is REQUIRED)
-            add("Implement OAuth", "Add OAuth2 login flow",
-                entity_type="task", project="sibyl-project", priority="high")
-
-            # Create a project
-            add("Auth System", "Authentication and authorization",
-                entity_type="project", repository_url="github.com/org/auth")
-        """
-        return await _add_mcp_entity(
-            title=title,
-            content=content,
-            entity_type=entity_type,
-            category=category,
-            languages=languages,
-            tags=tags,
-            related_to=related_to,
-            metadata=metadata,
-            project=project,
-            priority=priority,
-            assignees=assignees,
-            due_date=due_date,
-            technologies=technologies,
-            depends_on=depends_on,
-            repository_url=repository_url,
-            check_conflicts=check_conflicts,
-            skip_conflicts=skip_conflicts,
-            conflict_threshold=conflict_threshold,
-        )
-
-    # =========================================================================
-    # TOOL 10: remember
-    # =========================================================================
-
-    @mcp.tool()
-    async def remember(
-        title: str,
-        content: str,
-        kind: MemoryKind = "episode",
-        domain: str | None = None,
-        project: str | None = None,
-        tags: list[str] | None = None,
-        related_to: DeclaredRelatedTo = None,
-        task_ids: list[str] | None = None,
-        active_task: bool = True,
-        metadata: dict[str, Any] | None = None,
-        retrieval_keys: list[str] | None = None,
-        idempotency_key: str | None = None,
-        spans: list[dict[str, Any]] | None = None,
-        atomic: bool = False,
-        probes: list[str] | None = None,
-    ) -> dict[str, Any]:
-        """Remember durable context from planning, ideation, building, or any domain.
-
-        Use this aggressively during agent work to capture decisions, plans,
-        ideas, claims, procedures, artifacts, sessions, and domain facts. This
-        is the capture companion to the context tool: context retrieves what
-        matters, remember stores what future agents should not have to relearn.
-        Provide task_ids for exact task context. With a project, active_task
-        links the memory to the single active doing task when one exists.
-
-        Use retrieval_keys for the exact strings a future agent will search by
-        and that similarity cannot reach: an error code, an env var, a symbol,
-        a commit SHA, a config flag, an alias. You know them now and the reader
-        will not, so declare them even when the body already mentions them, and
-        especially when it does not. A query containing one of these strings
-        matches this memory exactly, case-insensitively. Up to 16 keys, 200
-        characters each. Skip keys for a memory nobody would look up by name.
-
-        Say what this memory does to the ones it relates to. A bare ID in
-        related_to links untyped, but prefixing an ID with a predicate is read
-        by retrieval when it walks the graph, and the declared predicate is
-        what decides how strongly a neighbor gets pulled in:
-        related_to=["supersedes:ep_0a1b"] says this memory replaces that one,
-        and the same shape works for "contradicts:", "requires:",
-        "supports:" (this is evidence for that), and "decides:" (this settles
-        that). The memory you are writing is always the subject, so read it as
-        "this new memory supersedes ep_0a1b". Nothing else is a predicate: any
-        other prefix is treated as part of the ID and links untyped.
-
-        You can describe the shape of what you are storing, and you know it
-        better than any cutter reading it afterwards.
-
-        spans: where this memory's sections begin and end, as
-            [{"start": 0, "end": 812, "label": "Root cause"}, ...]. Offsets are
-            character positions into content after stripping leading and
-            trailing whitespace, half-open so one span's end is the next one's
-            start. They must cover the whole body with no gap and no overlap,
-            first starting at 0 and last ending at the final character. Each
-            span becomes an independently retrievable passage holding exactly
-            that text; labels are indexed with it, so name the section the way
-            someone would ask for it. Without spans, long memories are cut by a
-            mechanical prose cutter that guesses at the seams. An invalid plan is
-            rejected with the offset that broke it, never silently ignored.
-        atomic: true when the body is one thing that must not be cut, such as a
-            single fact, a short rule, or a snippet that only reads whole. Not
-            combinable with spans.
-        probes: up to 5 questions this memory has to answer later, for example
-            ["why did the JWT refresh fail silently"]. Each is run through the
-            live search path the moment the write lands, and the response reports
-            the rank it came back at or that it did not come back at all. Use
-            them when it matters that a memory is findable and not merely stored:
-            an absent probe is telling you to rewrite the memory now, while you
-            still have the context, rather than discovering months later that
-            nothing could reach it.
-        """
-
-        return await _remember_mcp_memory(
-            title=title,
-            content=content,
-            kind=kind,
-            domain=domain,
-            project=project,
-            tags=tags,
-            related_to=related_to,
-            task_ids=task_ids,
-            active_task=active_task,
-            metadata=metadata,
-            retrieval_keys=retrieval_keys,
-            idempotency_key=idempotency_key,
-            spans=spans,
-            atomic=atomic,
-            probes=probes,
-        )
-
-    # =========================================================================
-    # TOOL 11: reflect
-    # =========================================================================
-
-    @mcp.tool()
-    async def reflect(
-        content: str,
-        source_title: str = "Session reflection",
-        intent: Literal[
-            "build", "plan", "ideate", "research", "debug", "decide", "learn", "general"
-        ] = "general",
-        domain: str | None = None,
-        project: str | None = None,
-        related_to: list[str] | None = None,
-        task_ids: list[str] | None = None,
-        active_task: bool = True,
-        persist: bool = False,
-        persist_source: bool = True,
-        persist_review: bool = False,
-        cited_ids: list[str] | str | None = None,
-        limit: int = 12,
-    ) -> dict[str, Any]:
-        """Reflect raw notes into reviewable durable memory candidates.
-
-        Use this after planning, ideation, debugging, or building sessions to
-        extract decisions, plans, ideas, claims, artifacts, procedures, and
-        session checkpoints. Set persist=True when the candidates should be
-        written back into Sibyl. Set persist_review=True to store them in the
-        raw review queue instead of graph promotion. Provide task_ids for exact
-        task context. With persist=True and a project, active_task links
-        persisted output to the single active doing task when one exists. Use
-        cited_ids for context/search IDs that materially informed the reflection.
-        """
-        return await _reflect_mcp_memory(
-            content=content,
-            source_title=source_title,
-            intent=intent,
-            domain=domain,
-            project=project,
-            related_to=related_to,
-            task_ids=task_ids,
-            active_task=active_task,
-            persist=persist,
-            persist_source=persist_source,
-            persist_review=persist_review,
-            cited_ids=cited_ids,
-            limit=limit,
-        )
-
-    # =========================================================================
-    # TOOL 12: manage
-    # =========================================================================
-
-    @mcp.tool()
-    async def manage(
-        action: str,
-        entity_id: str | None = None,
-        data: dict[str, Any] | None = None,
-    ) -> dict[str, Any]:
-        """Manage operations that modify state in the knowledge graph.
-
-        The manage() tool handles all state-changing operations including task
-        and epic workflow, source operations, and analysis.
-
-        Task Workflow Actions:
-            - start_task: Begin work on a task (sets status to 'doing')
-            - block_task: Mark task as blocked (data.reason required)
-            - unblock_task: Remove blocked status, resume work
-            - submit_review: Submit for code review (sets status to 'review')
-            - complete_task: Mark done (data.learnings optional)
-              Include data.cited_ids to stamp memories that informed completion.
-            - archive_task: Archive without completing
-            - update_task: Update task fields (data contains updates)
-            - add_note: Append a note to a task (data.content required)
-
-        Epic Workflow Actions:
-            - start_epic: Begin work on an epic
-            - complete_epic: Mark an epic complete (data.learnings optional)
-            - archive_epic: Archive an epic (data.reason optional)
-            - update_epic: Update epic fields (data contains updates)
-
-        Source Operations:
-            - crawl: Trigger crawl of URL (data.url required, data.depth optional)
-            - sync: Re-crawl existing source (entity_id = source ID)
-            - refresh: Sync all sources
-            - link_graph: Link document chunks to knowledge graph (entity_id = source ID, optional)
-            - link_graph_status: Get status of pending graph linking
-            - correct_memory: Correct a raw memory (entity_id = raw memory ID,
-              data.action = wrong, stale, duplicate, superseded, or revise)
-
-        Analysis Actions:
-            - estimate: Estimate task effort from similar completed tasks
-            - prioritize: Get smart task ordering for project
-            - detect_cycles: Find circular dependencies in project
-            - suggest: Get knowledge suggestions for a task
-
-        Server health and graph statistics are exposed as the sibyl://health and
-        sibyl://stats MCP resources, not as manage() actions.
-
-        Args:
-            action: Action to perform (see categories above)
-            entity_id: Target entity ID (required for most actions)
-            data: Action-specific data dict
-
-        Returns:
-            Result with success, action, entity_id, message, and data
-
-        Examples:
-            manage("start_task", entity_id="task-123")
-            manage("complete_task", entity_id="task-123",
-                   data={"learnings": "OAuth needs exact redirect URIs"})
-            manage("crawl", data={"url": "https://docs.example.com", "depth": 3})
-            manage("link_graph")  # Link all pending chunks
-            manage("link_graph", entity_id="source-123")  # Link specific source
-            manage("link_graph_status")  # Check pending work
-            manage("correct_memory", entity_id="raw-123",
-                   data={"action": "wrong", "reason": "Contradicted by source"})
-            manage("estimate", entity_id="task-456")
-            manage("add_note", entity_id="task-123", data={"content": "Root cause found"})
-        """
-        return await _manage_mcp_action(
-            action=action,
-            entity_id=entity_id,
-            data=data,
-        )
-
-    # =========================================================================
-    # TOOL 13: logs (Developer Introspection)
-    # =========================================================================
-
-    @mcp.tool()
-    async def logs(
-        limit: int = 50,
-        service: str | None = None,
-        level: str | None = None,
-    ) -> list[dict[str, Any]]:
-        """Get recent server logs for debugging and development.
-
-        Returns log entries from the server's in-memory ring buffer.
-        Useful for debugging issues without needing direct server access.
-
-        Requires OWNER role (super admin equivalent).
-
-        Args:
-            limit: Maximum entries to return (default 50, max 500)
-            service: Filter by service name (api, worker)
-            level: Filter by log level (debug, info, warning, error)
-
-        Returns:
-            List of log entries with timestamp, service, level, event, context
-
-        Examples:
-            logs()                    # Last 50 entries
-            logs(limit=100)           # Last 100 entries
-            logs(service="worker")    # Worker logs only
-            logs(level="error")       # Errors only
-        """
-        from sibyl_core.logging import LogBuffer
-
-        # Require auth context
-        ctx = await _require_mcp_context()
-
-        # Check OWNER role (super admin)
-        await _require_owner_mcp_context(ctx)
-
-        # Clamp limit
-        limit = min(max(1, limit), 500)
-
-        # Get logs from buffer
-        buffer = LogBuffer.get()
-        entries = buffer.tail(n=limit, service=service, level=level)
-        return [e.to_dict() for e in entries]
 
 
 def _register_resources(mcp: MCPServer) -> None:
@@ -2929,7 +96,7 @@ def _register_resources(mcp: MCPServer) -> None:
         from sibyl_core.tools.core import get_health
 
         # Get org context (optional for health - basic health works without org)
-        org_id = await _optional_mcp_org_id()
+        org_id = await mcp_context.optional_org_id()
         health = await get_health(organization_id=org_id)
         return json.dumps(health, indent=2)
 
@@ -2950,6 +117,6 @@ def _register_resources(mcp: MCPServer) -> None:
         from sibyl.persistence.graph_runtime import get_graph_stats_payload
 
         # Get org context (required for stats)
-        org_id = await _require_org_id()
+        org_id = await mcp_context.require_org_id()
         stats = await get_graph_stats_payload(org_id)
         return json.dumps(stats, indent=2)
